pyproject.toml
[project]
name = "nonebot-plugin-dailynews"
version = "0.1.0"
description = "NoneBot2 每日新闻插件，定时发送到QQ群"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nonebot2>=2.0.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "apscheduler>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/nonebot-plugin-dailynews"

[tool.nonebot]
plugins = ["nonebot_plugin_dailynews"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "nonebot_plugin_dailynews"}

[tool.setuptools.packages.find]
where = ["."]
include = ["nonebot_plugin_dailynews"]
